KNOWLEDGEGRAPH/requirements.txt
pdfplumber
decimer
rdkit-pypi
networkx
pandas
opencv-python-headless
tensorflow
pydantic